Migrate to sbt.AutoPlugin
