+++ tests/src/sbt-test/tut/test-04-only/build.sbt
@@ -1,4 +1,4 @@
-tutSettings
+enablePlugins(TutPlugin)
 
 scalaVersion := sys.props("scala.version")
